requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
telethon>=1.34,<2.0
PySocks>=1.7,<2.0
